renamed TerminalLimits to OperationalLimits; simplified UNION in query; reduced complexity of conversion method
