+++ cgmes/cgmes-model/src/main/java/com/powsybl/cgmes/model/CgmesModel.java
@@ -56,7 +56,7 @@ public interface CgmesModel {
 
     PropertyBags terminalsCN();
 
-    PropertyBags terminalLimits();
+    PropertyBags operationalLimits();
 
     PropertyBags connectivityNodes();
