+++ cgmes/cgmes-model/src/test/java/com/powsybl/cgmes/model/test/FakeCgmesModel.java
@@ -25,7 +25,7 @@ public final class FakeCgmesModel implements CgmesModel {
     private PropertyBags substations;
     private PropertyBags voltageLevels;
     private PropertyBags terminals;
-    private PropertyBags terminalLimits;
+    private PropertyBags operationalLimits;
     private PropertyBags connectivityNodes;
     private PropertyBags topologicalNodes;
     private PropertyBags switches;
@@ -59,7 +59,7 @@ public final class FakeCgmesModel implements CgmesModel {
         substations = new PropertyBags();
         voltageLevels = new PropertyBags();
         terminals = new PropertyBags();
-        terminalLimits = new PropertyBags();
+        operationalLimits = new PropertyBags();
         connectivityNodes = new PropertyBags();
         topologicalNodes = new PropertyBags();
         switches = new PropertyBags();
@@ -120,8 +120,8 @@ public final class FakeCgmesModel implements CgmesModel {
         return this;
     }
 
-    public FakeCgmesModel terminalLimits(String... ids) {
-        fakeObjectsFromIdentifiers("OperationalLimit", ids, terminalLimits);
+    public FakeCgmesModel operationalLimits(String... ids) {
+        fakeObjectsFromIdentifiers("OperationalLimit", ids, operationalLimits);
         return this;
     }
 
@@ -292,8 +292,8 @@ public final class FakeCgmesModel implements CgmesModel {
     }
 
     @Override
-    public PropertyBags terminalLimits() {
-        return terminalLimits;
+    public PropertyBags operationalLimits() {
+        return operationalLimits;
     }
 
     @Override
